Recover from the /rw read-only issue.
If the UBIFS failed to recover from LEB scanning, the mounting will fail. The
LEB scanning happens when ubi volume is mounted. So if the mounting fails, we
should try to reformat /user partition to recover from failure.

The mounting failure could cause a lot of programs that running on rw space to fail.

Change-Id: I3f4f0e939debae850845bb20b6a4bcb3edddd150