upload-crash-log: upload log incrementally.
By default, only upload the dmesg content since the last successful upload.
This might even include the messages since the last reboot, unless they've
already been uploaded.

Right now the server side isn't smart enough to merge all the incremental
uploads back together, but this provides enough information to do it.  For
now it'll appear as a bunch of separate log files.

This adds /etc/init.d/*uploadlog, which runs upload-crash-log in a loop for
as long as the system is running.

Change-Id: I6c7daa6338a227fe8fb907a83fa4da6eaf5e39be