logupload/server: better handling for incremental uploads.
Now an old-style "File" object still consists of a single upload, but an
upload is assumed to be a timestamped chunk of a log file.  All the files
for a given machine id should be concatenated to form the entire log.  We
add a new Machine object to arrange them all under.

The toplevel view is now a list of Machine objects; click on one and you can
get a list of the File objects available.  Picking a File gives you all the
log messages from that file *onward*.

Someday this should probably be refined to only give you max one day's worth
of logs, or even let you fine tune the times more precisely than that.
(Actually you can by manually editing the URL and providing ?start= and
&end= parameters to define the time period.  But there's no UI for it yet.)

We also completely remove the fulltext search feature; it didn't work right
anyway because of appengine suck.  Someday appengine will release an actual
fulltext search feature, and then maybe we can have it.  For now, you search
the hard way (by opening log files and using your browser's search-on-page
feature :))

Change-Id: I8b381f2f27d7002d38ce36657490939136a6fdcd