platform/bruno changes for simpleramfs.
These changes are needed in order to work with the simpleramfs changes in
buildroot.  Since simpleramfs has its own startup script, all the stuff in
/etc/init.d now needs to run, not just the stuff that used to run in the
initramfs.  Also, the feature of S04volcheck that used to mount root can go
away; now we only mount the non-root volumes.  And we can remove some
scripts that are no longer called.

Change-Id: I38de447f79fa75dad500bba5fff9fb3ddf9ee0ac