gfiber-dropbox: update to use ndb and more caching.
Also, approximately split up logs from one reboot to the next, which makes
it more likely that you won't be downloading zillions of megabytes of stuff
when you click on an older log link.  (The splitting probably still needs
more work, like perhaps a hard limit on the log size, splitting it into a
new chunk when the limit is exceeded.  But so far I haven't seen that
problem, so this works for now.)

With ndb plus lots of caching and parallel fetches, performance is now very
nearly tolerable.  Nearly.

Change-Id: I74f985923a6a668ea2a26e32301f6d819f6c36aa