klogd: start it in the background.
klogd uses a weird method of daemonizing itself, in which it forks into the
background and then tries to send its parent self SIGTERM.  This can have
obscure problems in cases where getuid() != geteuid(), which basically
causes it to hang during startup.

To avoid that kind of hang, let's just put an & at the end.  We shouldn't do
that with syslogd since it's important for syslogd to *really* be running
before any syslog-using services start, but for klogd it should be harmless.

Change-Id: I36829b36161db65c72f9c2d72948a22e9ecf2a6e