upload-logs: abort if it takes more than 60 seconds to upload.
If that happens, it might be because the network was down or our TCP session
got silently interrupted for some reason.  (Because of the nature of TCP, if
we are stuck in recv() we might never get notification that the other end
has died.)

Apparently some people have observed this happening in the field, causing
uploads to sometimes get permanently stuck.

Change-Id: I2df753a32f4538acd8888af46381ac91f173d0f5