dhclient-script: more careful /tmp/cwmp_acs_url management.
Including more careful persisting of values across reboots, and a hardcoded
fallback URL if all else fails.

We also now replace the file atomically, to prevent any possibility of
leaving the filesystem with an empty file.  As a bonus, this will fix up any
instances of a bug someone found where /config/tr69/acs_url was a circular
symlink back to /tmp/cwmp/acs_url.  (mv'ing a file onto a symlink replaces
the symlink with the file.)

Change-Id: I3295bcb1f6ae3a4e521b9c8488493cad61c21864