stresstest: reduce /dev/mtd read size from 64k to 4k.
It seems read(65536) actually makes the kernel try to allocate 64k of
contiguous kernel memory, possibly in an interrupt context, which can return
an error.  I don't see any good reason the mtd driver should have to do that
(why not just do it in smaller batches?) but other people on the Internet
seem to have reported the same problem.  This will make the problem go away
in stresstest.

Note that there's a possibility that this memory allocation problem is
related to other nand-related problems.  If so, we should see messages of
the form "page allocation failure. order=" in the dmesg log.  This won't fix
those, since it's only changing the stresstest.

Note that I've only ever seen this on fat bruno, not thin bruno, so it's
probably related to there being disk activity at the same time.

Change-Id: I28414d9b492a7e657b9627a7c2839f1941518c41