upload-crash-log: use dmesg -r
When we dropped busybox, we switched to the util-linux version of dmesg,
which has the unhelpful behaviour of discarding KERN_DEBUG log messages by
default.  Let's add the -r option to just dump the raw contents of the
kernel message buffer.  Otherwise we'll be missing a lot of the good stuff
(including almost all the userspace logs, which are logged at KERN_DEBUG!)

Change-Id: I8a748b1954d3f0493be3fd561bff6bcf4efde10f