logupload/server: part of the last optimization was too expensive.
Generating the list of available logs for a given box costs 1000 datastore
reads or so (which is, incidentally, one reason it's so slow).  This is sort
of okay (albeit slow) when you do it in response to a human loading the
page, but not okay when you do it every time any machine uploads a log,
which is once per machine per 5 minutes.

The right optimization is to memcache things in smaller chunks so the
1000-unit fetch will be needed almost never.

Change-Id: Id63f340333555ed6005f4d3bba036af6e4decc9d