gpio-mailbox: earlier race condition fix added a race condition.
Even if the reset button was pressed at startup, the reset_button_msecs file
would only get created the *second* time through the loop (after the 'ready'
file was created), because the default value of int reset_amt was 0, so
write_file_int didn't write it as long as it was 0.

This would normally have been fine, since we don't even want to create the
file if the button isn't pressed.  But the *first* time through the loop
when we notice the button is pressed, we set reset_start = now, so the
amount of time it has been pressed is 0, even though it's pressed, and so
the file wouldn't be created.

This bug has been there all along.  But previously, we did the fan tick
polling loop and *then* wrote the ready file.  S05factoryreset wakes up
instantly (thanks to inotify) upon the ready file being created, then looks
for the reset_button_msecs file.  It was still a race, but before,
gpio-mailbox looped again instantly, so within a few microseconds the
reset_button_msecs would be updated.  Now, there's a delay right after
creating the ready file, giving S05factoryreset time to swoop in, see that
the reset button isn't pressed, and move on.

Interestingly, this happens only on flash, not nfsroot, because of the
particularities of the timing.

This patch double fixes it: we set the default value of reset_amt to -42
(like the other integers) so it will be written even if 0.  We also set
reset_start to "now - 1", so that we never try to write a 0 to the file,
just in case someone someday writes code that assumes it isn't pressed if
the time it has been pressed is 0.

Change-Id: I8af2389d4c7a8225a576938196f12c0eaf7ac522