ginstall: earlier fix to IsIdentical broke kernel installs on B0.
Always seeking back to offset 0 isn't really safe.  But let's throw an
exception if someone misuses IsIdentical.  This would have instantly caught
the original bug that the IsIdentical change was intended to fix.

(This is the same as a similar fix on the jaguar branch, but also updates
the ginstall_test.py.)

Change-Id: Iabf4dd2b018a3ac96fee5f0067c68f6672d26f17