Update ginstall for SLC support.
The squash_ubi format will still be supported in ginstall. If the rootfs
is in squash_ubi format, the existing logic is executed. If the rootfs
is in squash format, we will attach the corresponding rootfs partition,
install the rootfs and let the kernel handle insertion of UBI
headers. The drawback is we will do verification like installing to
other MTD partitions. But since installation is not time-critical, this
is not really an issue.

Change-Id: I4a181d08be224e493817f2e92bae14621a060067