Add readallfiles script to read through the rootfs.
This is to force verity check for all files in rootfs and set failure
count back to zero. It also serves the purpose to prefetch all rootfs
contents to linux cache. The overall time for scanning the whole rootfs
takes about 50 seconds when the process runs at the "nicest" level.

Change-Id: I013fd0c4a37cdfbe7fbb8f7ef69a29eb29bb0d7a