gpio-mailbox: override fanpercent to 100% if CPU temp >= 100 degrees.
Bugs in sysmgr caused it to freeze and fail to set the fanpercent in some
cases.  That could lead to critical overheating, so let's override it when
things get really bad.  This is a pretty blunt instrument, but it's better
than melting down the box.

While we're here, let's also print out the requested fanpercent in the
periodic output, to facilitate later diagnosis.

Change-Id: I1e0cabfc8677d16c571260b5a2d688c6e7eaa833