gpio-mailbox: create /tmp/gpio/leds-ready after updating leds.
That gives a way for other processes to check that their led command has
been received: delete /tmp/gpio/leds-ready, then wait for it to be created
again.

Since the loop happens rather frequently, avoid churning the filesystem by
opening with O_EXCL.  That way if nobody deletes the file, it won't be
re-created (with associated inode updates, etc).

Change-Id: Ifb58091ba5b28e1400f0a5d499e0cd3c6ebfcf2e