Add packaging for options.py
The current image packaging process only keeps .pyc files. This CL is to
package options.py to python target directory.
There are two approached to package options.py, either adding a new
package in package/py or embed it in platform Makefile. After discussion
with Avery, we agree to maintain it in platform.

Change-Id: Ic7b8cb3c97b881c846bed012e4cdab815ec3ce10