ginstall: refuse to install old software versions on new-style partitions.
We check the size of the rootfs0 partition.  If it isn't old-style, then
refuse to install old-style software on it.

I have a longer term plan for better and safer all-around ginstall package
versioning, but this change should do the job for now, and it's small enough
that it's unlikely to introduce new bugs.

Change-Id: I82a8a3cf54df6892c8a4089b83d590dc60f75268