gpio-mailbox: suspend the child until parent has closed nexus.
This is mainly so that /var/run/gpio-mailbox doesn't get created until the
parent process has closed down its copy of nexus, so that other processes
can wait on that file and know when it's safe to start.

Nobody has observed this problem in real life, but it's possible, so let's
fix it now before it causes weird hard-to-debug problems.

Change-Id: If4ee24f156b756d6b78c1bbd6480f08417b1ec7d