Add script to generate signature for serial no.
The purpose is that CFE, in RMA recovery mode, will enable the USB
driver and look on the USB storage for a new CFE (presumably an
unlocked one) to install. However it will only install that CFE if
there is also a signature file on the USB that signs this particular
device serial number.

Change-Id: I40c942df53dbf324d15fa3f5e3ebfa2c2c24bb04