logos: use multiple token buckets to allow for short-term bursts.
The idea is that you sometimes need to write a lot of stuff at once - like
in a really big crash dump - but you won't do that very often.  So we want
to have separate limits for per-log-upload-cycle (don't overflow the dmesg
buffer in 300 seconds) and per-day (don't overflow the server-side log quota
in 24 hours).

To implement these limits, we use two token buckets, plus a third "helpful"
token bucket to let you know when you're getting too dangerous:

 - the "burst" bucket fills quickly but is small (<= kernel buffer size)
 - the "daily" bucket fills slowly but is big (<= daily quota)
 - the "warning" bucket fills slowly and is small (<= kernel buffer size)

If the "warning" bucket runs out of tokens, you get a message that your rate
is unsustainable; if this is an exceptional case (such as a crash dump) you
can ignore the warning.  Otherwise you should probably fix it, because
your program will end up having its logs messages rate limited within 24
hours.

Because we now have long-term rate limits, we also increase the assumed size
of the kernel log buffer from 1M to 10M.  Previously, we had to be careful
not to make that buffer too big, or we could end up uploading the full
amount every 300 seconds, resulting in super-huge space usage on the server.
We don't have to worry about that now, so increase the short-term buffer so
we don't lose bursty log messages on devices that are normally quiet.

Change-Id: I95666ca64fccc98ceac879d134aa5e7ebc597e81