diskbench: a tool for testing disk/cpu performance with real-time streams.
This program measures CPU usage and looks for buffer over/underruns when
running a configurable number of reader and writer streams to separate files
in parallel.

It also offers configurable settings to enable/disable O_DIRECT,
fallocate(), and ionice, as well as block sizes and data rates, to see what
difference they make to performance.

Change-Id: Ia177d1057a5ee2d9f71b925ffe6b2b1e7e41495c