Add 'sdbbp' command.
This just runs the assembly 'sdbbp' instruction on MIPS, which triggers a
debugger breakpoint.  We can use this to validate whether the debugger
breakpoints are implemented correctly, particularly when not debugging.

It's important to validate this up front, otherwise crashme could execute an
sdbbp instruction and crash the machine in a boring way instead of an
interesting way.

Change-Id: I3920674b3abe623eca37d01d1130603eed96671d