rtwatcher: remove strace stuff.
When we're low on memory and rtwatcher complains that no non-realtime
processes have run in the last few seconds, the strace activity seems to
make things even worse.  And so far it has never helped us debug a real
problem.  So let's leave it out.  The ps listing is more useful anyway, and
causes less trouble.

Change-Id: I0b7cbe2edc032c0f163f3f9e2204a8139cdac593