Add bigbin, an enormous binary
EXPERIMENTAL: Doesn't trigger any interesting problems.
I was hoping to reproduce TEXT corruption.

Looking for issues with paging and TEXT segments.
1. bigbin has a giant TEXT segment consisting of a
   long series of trivial add instructions.
2. Modify balloon to be able to allocate a set amount of
   memory, and optionally exit. This allows it to put the system
   under memory pressure, without killing the whole system.

bigbin takes nearly a minute to compile, due to the time spent
assembling. Though the resulting binary is enormous, it compresses
extremely well and doesn't significantly increase the image size.

It _does_ take a lot of memory when run. That is rather the point.

Change-Id: I10f90617276f90a05c34da91e0148363507c31df