8725917: Increase interval of HDD temperature check in sysmgr
Currently sysmgr checks the HDD temperature every 15secs, which is way
too often and needs to be changed to something like 300secs (5mins).
Additional complexity is coming from the fact that the thread checking
the HDD temperature also checks the SoC temperature and controls the
fan speeds. Those should stay on the old interval of 15secs.

The fix introduces a new option '--hdd_temp_interval' that defines a
new interval just for checking the HDD temperature. Since the thread
continues to run at the regular old interval (15secs), the hdd
interval should be a multiple of it. If it is not it will simply run
the next time the thread runs. Currently, the HDD check interval is
set to 300secs (5min).

Change-Id: I000f6e04c2bad8dc0367b96dfd2a7c42cf08c219