Improve soft remove control
This adds a macro that interprets a string of digits as individual
digit presses followed by the OK key.  E.g.: '304' will be translated
into 'DIGIT_3 DIGIT_0 DIGIT_4 OK'. This makes it easier to change
channels.

In addition, the logic flow was slightly changed to fit the new macro
better.

Change-Id: Ia21e800d0f70da431a8e7c19e8868dc3b0ef99b8