Remove custom build rules for logos and bigbin.
We don't need these, as there are already implicit rules that do the job.
logos was just missing the LIBS entry that makes it link with libstacktrace.

This fixes building for kvm, which uses glibc, which puts clock_gettime()
in -lrt, which was included in logos until this explicit rule was introduced
that forgot to pass $(LIBS) on the linker command line.

Change-Id: I94eea144178a067baf402a2c875fe6c540696913