9850181: alivemonitor should log when monitored process hasn't been heard from for a while
This changes the alivemonitor parameters as follows:

<keepalive_file> <first_check> <incr_checks> <timeout> <command>

The keepalive logic runs in cycles. A cycle begins and ends with a
successful check of the <keepalive_file>, i.e., it was touched since
the last cycle. The first check starts <first_check> secs after the
cycle begins. Incremental checks are done at <incr_checks> intervals,
until <keepalive_file> was found to be updated or <timeout> is
reached. In the former case, the cycle restarts, while in the latter
(timeout) case, the process is restarted and the cycle starts
again. At each checkpoint before the timeout is reached, a log entry
is created if the keepalive file hasn't been updated.

Change-Id: I347894efc11aeb5d3273ac5b9953ac345b8ba53b