soft_rc improvements
(1)
Soft_rc provides the option to read commands from an input file, but
if this file is a FIFO pipe, it reads exactly one time and then hits
EOF. This is fixed by opening the file as 'r+' instead of 'r'.

(2)
Errors are fed into the logs instead of printed, plus stdout is
flushed to make sure logs appears in the right order relative to other
logs.

Change-Id: If153f90751cb2d7c289e8bf9d64f0b166cdd7353