ginstall: install the loader on 7425B0 and 7429 devices.
The old code for skipping the loader on 7425B0 devices has not been needed
for a while; recent loaders (~12 months) all work fine.  To be extra
careful, bumped the minimum supported version to be a more recent one. This
allows us to remove the special check for 4GB devices, since we just don't
support any images anymore that were 4GB-only.

I also confirmed that there is no reason to strip the kernel signature on
7429 devices with our current bootloader; it works fine with the signature
left in place, so let's leave it in place.

Plus, change a few places to be more careful with parsing.  An empty list of
supported platforms doesn't mean "any" (which is super dangerous); it means
"none." With v2 files, we'll just hardcode the exact platforms that they
support.

Finally, don't crash on platforms where sgdisk is missing.  We need it for
finding partitions on platforms that have the system installed on disk
partitions, but on flash-based systems, it's not needed and we can just skip
it.

Change-Id: Ib114eea15e8f090a2d013972ab4ba0daf9eeaa8b