Remove sdbbp.c.
It isn't that useful since we now have a bootloader which doesn't crash on
this instruction.  And when we rebuild the gcc toolchain, it claims not to
recognize the instruction.  Which is probably a problem, but not a big one,
and the rest of the system builds, so let's just drop this for now.

Change-Id: Idc3e60ec6148418c461084c732a2dfa9382a2e96