Add tool to generate signatures for Mindspeed platform
The new tool mimics the existing brcm_sign_enc tool, but
generates a signature that we actually understand (necessary
because we have to write the verification code for this one.)

The signature is a standard RSA signing of a SHA1 hash, and
is computed by calling out to openssl.

Also added ability for repack.py to choose the tool it calls.
Defaults to brcm_sign_enc, overridden via -t flag.

Change-Id: Ieb6ffa3f2040a53e66e85d9b35b7c15ee7be05bf