ginstall: Reduce block size from 256kB to 64kB
When ginstall writes to flash, it only writes entire blocks of data. If
the length of the input file is not a multiple of the block size,
ginstall will pad it with 0xFF bytes at the end.

Previously, it used a block size of 256kB which is too large for the
uloader partition on Optimus (128kB). This resulted in the following
error message "IOError: [Errno 28] No space left on device" although the
uloader has been properly written to the NOR flash. Let's reduce the
block size to 64kB.

ginstall prints a dot (".") for every block it writes. As a side effect
of this change, it now prints almost or exactly four times the number of
blocks.

Change-Id: I1c822c6e51a2b76d4320dd046e20060d30048d62