alivemonitor: Avoid using POSIXLY_CORRECT env var
When using getopt, we want to stop option processing as soon as a
nonoption argument is encountered. Previously, we achieved this by
setting the environment variable POSIXLY_CORRECT, but this turned out to
be problematic as it changes the behavior of child processes in
unexpected ways.

The same behavior can be achieved by prefixing the optstring with "+".
So let's do this instead of setting POSIXLY_CORRECT.

Change-Id: Icc0a552f35d496bcf54f35ca44bd14d811acc79f