ginstall: Implement minimum_version attribute
If the manifest in the .gi file contains a minimum_version attribute,
then refuse to install if the running image is older than the minimum
supported software version specified in the manifest.

This can be helpful if changes need to be made to the image format that
would cause older version of ginstall to wreak havoc. However, this does
not stop versions of ginstall that are already in the field.

b/13550937

Change-Id: I08a111a73c04f910d22c7097c4f12af604b14c0e