ginstall: Implement image format version 4
Allow the manifest file to be called either "manifest" or "MANIFEST".
This allows us to deliberately break compatibility with older versions
of ginstall that potentially brick GFRG2xx units by installing incorrect
uloader versions.

Going forward, gi files for GFRG2xx platforms will have their manifest
file named "MANIFEST" instead of "manifest". This will cause older
versions of ginstall to abort the installation. Newer versions of
ginstall i.e. versions that contain this commit and commit

246b14a ginstall: Don't install unsigned uloader on secure optimus

will install the image just fine.

Also, check the version of the manifest. If it's not version 2, 3 or 4,
refuse to install the image. This allows us to break compatibility in a
cleaner way in the future if there is need. Previously, we had to revert
to hacks like renaming components in the gi file like renaming
"manifest" to "MANIFEST".

Change-Id: I6eceb997f5578e3ab04f5006a55437ec0849b98c