Use nandwrite/nanddump for bad block handling
Previously, we did not skip bad blocks when writing to NAND flash. As a
result, we were not able to use NAND flash chips with bad blocks in the
kernel partitions. Let's use nandwrite and nanddump which are bad block
aware.

As of today, CFE, the bootloader that is used on Broadcom CPE 1.0 devices,
cannot properly handle bad blocks. Let's not use nandwrite on these platforms.
Instead, we keep on using the old method that throws an error if it encounters
a bad block.

Change-Id: I4aeb4eed84f2a0f19009bcdccdbacbacd66d806d