waveguide: first version of a daemon to coordinate wifi activity.
The overall idea is that waveguide runs as a daemon on all GFiber-managed
wifi access points.  Each waveguide instance on a given LAN will join the
waveguide multicast group and send periodic updates to all the other
waveguides on the same LAN segment.

By collecting the status updates, the instances can form a consensus about
the overall arrangement of the network and learn about the whole wifi
neighbourhood (including non-managed APs and APs on other nearby networks).
Using this consensus, each waveguide instance should be able to make a
decision about whether or not it should be enabled, whether it should use
2.4 GHz or 5 GHz or (if available) both, what power levels it should use,
and whether it should encourage certain stations to roam to another AP in
the waveguide group.

Currently it doesn't do any of those things.  It does send and receive
packets on a multicast group though, and the fields in the packets it sends
are mostly filled in.

Change-Id: I0090331f99b00528c6e29763ebf84f3c55807e55