Write to failure count reg only on BCM platform
When S99readallfiles determined that the system booted up all the way
and that all files in the rootfs can be read, it writes to the register
AON_CTRL_SYSTEM_DATA_RAM[127] to communicate this fact to the bootloader
(CFE). This change ensures that this only happens on the Broadcom
platforms. Previously, it also tried to access this register on the
Freescale platform which resulted in memory corruption.

Change-Id: Ie02688dc40323d6a8e40d9546012595873b2c3a3