waveguide: implement auto-disable and some daemon tests.
When a waveguide-enabled AP with --high-power is detected nearby and the
current waveguide is not --high-power, write a file called
/tmp/waveguide/wlanXXX.disabled to indicate that the local wifi should be
disabled.  (waveguide does not actually disable the wifi interface; that's
someone else's job.)

While we're here, fix some portability problems, make things more
configurable (it now runs, in fake mode anyway, on MacOS), and add a
a daemon-level end-to-end test script, using a fake version of the
'iw' command.

Change-Id: I09ed9d261af95ae049e2cc73666b08c4425adc29