isoping: fix ttl setting and make it configurable.
Some people want to run isoping over the internet, so make the ttl field
settable on the command line.

Also, the way we were setting it returned EINVAL on MacOS X, because we were
setting the IPv4 TTL on an IPv6 socket.  On Linux they're separately
configurable, so set both.

Change-Id: Ie685edc5d6051d53f0f8c6ff4aa98e6bb9b4eafc