waveguide: remove low-priority scanning flag.
At some point something changed in the ath10k driver that made it not
support this flag anymore.  It might be just a firmware version problem, but
I don't think this flag does anything useful anyway, so let's remove it.

May have contributed to b/18432293 since it seems to affect how often the
scan actually succeeds.

Change-Id: I7aea679748c212b1d9ec5eef86ce1069c88acbcc