ginstall: Use hashes from manifest for verification
ginstall reads back the images it wrote to flash memory. Previously, it
performed a binary comparison with the images in the .gi file and did
not make use of the SHA-1 hash values that are part of the manifest.
With this change, ginstall reads back the data from flash, calculates
the SHA-1 hash value and compares it against the value in the manifest.
If the SHA-1 is not available for some reason, we fall back to the old
method.

Change-Id: Ia6c0f49d9b6ecc9e01294304d0a37d67aecb653e