wifiblaster: first iteration of wifi blast tester.
In this iteration, we take a simple approach that allows us to measure
throughput and loss without processing 802.11 acknowledgements.

We read the cfg80211 counters for tx_packets, tx_retries, and tx_failed
before and after a pktgen packet blast, taking the difference as our
measurement. Because this test is not run in a controlled environment,
the counters can include traffic outside of this test, but they are
assumed to be in the noise.

Throughput is computed as "tx_packets * size / time". This is accurate
as long as tx_packets represents the number of acknowledged packets.

Change-Id: I1e6e9a429b80098046e84a12acc226adb6bf8dea