ibeacon: remove interval, add enable.
Once set, Blue/Z will continue to advertise the data
it has been given. So ibeacon doesn't need to stay running,
it can set the advertising data and then exit.

Remove the remnants of the -i argument. We removed the
ability to set the BTLE advertising interval but didn't
remove the command line argument.

Add code to enable advertising, and to set the advertising
interval. We only want BTLE advertisements enabled when we
have a service to advertise, might as well enable them
along with the service.

Change-Id: I4d948ffa31ca5b82d553cbe8e151cfd127cc42bc