Fix a WindCharger Ethernet port diags stat problem
With the WindCharger kernel, the ifstat uses only
an int counter to show statistics. However, when the
counter (32 bit) overflow, it generates an incorrect
64 bit number where the upper 32 bits are meaningless
while the lower 32 bit became negative. To avoid showing
incorrect Ethernet traffic statistics in diags, this
problem is fixed by taking only the lower 32 bits of
data and negating it if it is negative (overflow).

Change-Id: If4e3198fe9fe5fba838fe78f88b6a5ff7368dd6d