WindCharger diags: add carrier monitoring for Ethernet port tests
1. Add network port carrier monitoring to detect cable
disconnect.
2. Make eth0 to eth1 bi-directional test more robust. Because
the test is performed with two processes, the processes have to
be stopped before querying the result. Since the processes are
stopped one after another. So the TX bytes sent could be a little
bit larger than the RX bytes received. Add a 1% margin in the
pass criteria for this test.

Change-Id: I9ea031350ec286c0aad6bd90222dfa2bf56447ff