Fixed b/19967333 coredumping
We had a mistake before where the program would not exit
if it failed to upload, that is now fixed. That will resolve
the core dump most likely because it was happening after
an upload failure occurred. I also noticed that when you had
an upload failure it would call curl_easy_cleanup on the handle
twice...which is illegal according to the API, so that may
have then caused a problem when curl was initialized again
for the upload after the failure.

Change-Id: If3799e818e893f1da4b278fb62b3e7f15ddb5304