freemegs: measure available memory.
Allocate memory and ensure that physical pages are allocated to it.
Also ensure that those pages do not compress well, to provide a test
for zram usefulness.

Print out each megabyte as it is allocated, and exit if it takes
too long. Even so, it often gets killed by the kernel's OOM Killer.

Change-Id: I1e44f69edc9e8632c3650b70e322fe9347f8b810