Fixed usage of timeouts in curl
We don't need to set NOSIGNAL, the prior crashes that
I thought were caused by this actually were not. That
crash was fixed in a later CL. This also adds an actual
timeout to curl. We no longer use alarm() because curl
uses that internally so ours didn't matter anyways (unless
we disable signals, then it won't use that, but then it could
take forever to timeout).

Change-Id: Ifcf746dd687486bf5ed743fdc9fb332d0041bbfb