/bin/wifi:  Fix slow/flaky autochannel (b/20491802).
There is a loop which tries to run 'iw dev <ifc> survey dump' up to 9
times, until it suceeds, with a 0.5s delay between each.  The code was
incorrectly look for this subprocess call to fail rather than succeed,
causing all nine attempts + delays to occur when nothing was wrong.

Additionally, if the first attempt failed, no retries would be
attempted.  The same change fixes this as well.

Change-Id: I4f8af5eeb8245147508c74ca24b54d07c2d4da2b