ginstall: remove --kernel and --rootfs
ginstall can install the kernel and rootfs from individual files.
This was especially useful way back at the beginning of the
project, but that was years ago. Now, its mostly a trap for the
unwary who would probably be better off using the *.gi image.

These options will also be very hard to support in the upcoming
stream-directly-to-flash support. So remove them.

Change-Id: I98ce58fd90101882921070bb0bbe8057dcb966a4