ginstall: stream directly to flash
--tar can now handle a URL, to download an image
and stream it directly to flash, without requiring
that the entire image be buffered.

The loader and uloader have to be buffered in RAM
before installation. They have a signature which
needs to be checked before installing.

The kernel and rootfs are streamed directly into flash.

Image formats without end-to-end SHA checksums, like
ginstall v2, had to change. Previously it would read
back and compare, but now we cannot seek() to read back
the image file again. Instead we calculate a SHA checksum
while the file is downloading, and then read back the
flash to calculate a checksum again to compare.

Change-Id: Ie8771166a6c98ab89250b290a051414be363feac