taxonomy: Use DHCP hostname, add signatures.
If the device sent a hostname in its DHCP request,
include it in a first signature lookup. Some Wifi
signatures are too undifferentiated to include by
themselves, but become much more distinct with the
hostname.

Don't special case 802.11a/b/g any more. With the
hostname, we'll be able to identify some of these
devices.

Add signatures for:
+ Playstation 4
+ Wii-U
+ XBox One
+ Chromecast
+ iPod touch 1st gen
+ iPod touch 3rd gen
+ Add another Nexus 7 (2013) signature

Change-Id: I72c660d84a2e766bdb033ec92aa24e683fc7e53c