isostream: expose option to limit connections
isostream has a compiled-in limit (#define MAX_CHILDREN) to the number
of connections to serve.

Expose this as a command-line option, to allow a particular instance of
isostream to serve fewer connections than it otherwise could.  This is
required by the "what if?" wireless performance experimentation plan, to
serialize requests from multiple prospective clients when we only want
one on the network.

Change-Id: I1861294fa2655c690908dcf7218b01a28d905005