optimus i2c: Use more generic sysfs path
The sysfs path /sys/devices/platform/comcerto_i2c.0/i2c-0/0-004c/ is not
valid on the new 4.1 kernel for Optimus.

Use the more generic path /sys/class/hwmon/hwmon0/. The only problem
with this approach is that, on the old kernel, the path is slightly
different /sys/class/hwmon/hwmon0/device/ (an extra "device/" at the
end). Make gpio-mailbox probe for the correct directory.

Change-Id: I06c0d44bff8159f881935a09590a74be619a0c3e