wifi: force restart of hostapd/wpa_supplicant if experiments change.
Previously, we'd announce that an experiment was successfully applied even
if we then chose not to restart hostapd because the config file hadn't
changed.  This could result in inaccurate experiment reporting.  To work
around it, add a comment to the config file listing which relevant
experiments were active at config time.

Also clean up some formatting while we're here.  It's a minor style thing,
but usually people use lists for argv rather than tuples.

Change-Id: I3456f996f8e54e60327e2f4b0e1d1d1c83cc9df2