wifiblaster: use python -S, and remove default --interface.
"python -S" is a cheap optimization to save a bit of startup time and
memory when we don't use many external modules.

The -i parameter defaulted to wlan0, which was misleading; calling
wifiblaster without a -i option might make you think it's scanning
all interfaces, when in fact it just defaulted to the first one.
Remove the default altogether.  Waveguide always passes a -i option
anyway.

Change-Id: Ifd9f1e9e2305d6d196229a3ddd3543dfa16c180c