wifi2json: transmogrify into wifi_files.
catawampus relies on wifi_files.py to write JSON files in /tmp/stations.
wifi_files.py is using about 9% of the CPU on GFHD200 and GFRG2x0,
which is impacting other processes (like TV). The overhead of
parsing the iw output appears to be the culprit.

Revamp wifi2json to match the format of wifi_files.py, and rename it
wifi_files. The C version of wifi_files uses negligible CPU time
on GFHD200 / GFRG2x0.

Change-Id: Iedf6a215bbe3cb54f5e2275e82df1a004d8f0931