ginstall: Add lockfile to ginstall.
Now if ginstall is running, it will create a lockfile to ensure it is
the only instance running.

When a potentially orphaned ginstall instance runs to completion, it
will create a status file, which can then be checked to verify that
ginstall has, at some other point in time, run to completion.  It is the
parent process's responsibility to decide what to do when encountering
this file.

b/22850099

Change-Id: I7e7ac77dca245c498b670fdaa08e85eaa055bb33