waveguide: make rapid polling use time.time.
rapidpolling used a pre-existing 'now' variable, which
is monotomic time (i.e. uptime) in the real system.
monotime isn't very nice for the user of the rapid
polling feature, as they have to read the uptime first
and then calculate the rapidpollinguntil value.

The unit test worked, but in the real system it didn't
do what we want. Make rapid polling use time.time()
instead.

Change-Id: I64af018e21327e9c1a2df87253dd9363f35639b0