ginstall: Fixes installation completion logic.
In the event of an error setting something in hnvram, HNVRAM_ERROR is
returned, but if it succeeds, the completion file (used to determine if
ginstall has run to completion) doesn't get written even though an
installation has completed successfully.

Also adds a bit of debug text (noticed it was hard to see what was going
on during testing).

b/23598573

Change-Id: Id577e21ac6c5539d91acefd116ff7ff417077e3c