wifiblaster controller: keep only the most recent test for each client.
We thought it would be a good idea to keep all of them, but it turns out the
extra data took too much space in cwmpd, which would then get forwarded
repeatedly to the ACS.  Periodic stats can capture the history instead.

While we're here, refactor the unit test to avoid unnecessary mocks.  As a
bonus, this also tests more of the code more thoroughly, and takes fewer
lines of code.

Change-Id: I39ed76ed44643e1de47e66810dc3fa6a38ddadf9