logupload/client: better handling of uncompressible zlib data.
Previously, logupload would silently abort if ever asked to upload data that
wasn't compressible.  Fix this, and add some tests to validate that it
works.  This was easily reproducible with:

   ./upload-crash-log2 --stdin foo </dev/null

Also, in stdin mode, we were unnecessarily terminating the buffer with a NUL
character as if we presumed it to be ASCII; don't presume that.  As a bonus,
don't overrun the buffer if it was already full when we appended the NUL
terminator.  In --stdout mode, use fwrite() instead of printf() to allow for
binary data.  Add a unit test that validates --stdin and --stdout mode work
as expected.

Change-Id: Ic02f952d36ded4d566db9d05356d73a3fc5ce725