logupload/client: add --freq parameter and allow --stdin to be incremental.
You can now specify --freq to change the logupload frequency.  Default value
is the same as before, an average of 60 seconds.  Zero means exit after the
first upload.

The --stdin option now loops by default (although it'll still exit if it
gets EOF).  If --freq is nonzero, it'll stop reading stdin and do an upload
as soon as the given timeout has expired, then resume reading stdin.  This
could be useful for uploading the output of a program directly, rather than
through kmsg.

Change-Id: I4798a59486b7522ff08e5ec8dc6055dcb50a7f32